+++ presentation/src/main/kotlin/team/sungbinland/sseukssak/activity/empty/Event.kt
@@ -1,0 +1,12 @@
+/*
+ * SseukSsak © 2022 Team Sungbinland. all rights reserved.
+ * SseukSsak license is under the MIT.
+ *
+ * Please see: https://github.com/sungbinland/sseukssak/blob/main/LICENSE.
+ */
+
+package team.sungbinland.sseukssak.activity.empty
+
+sealed class Event {
+    object OpenDrawer : Event()
+}
